bitmap: Use size_t instead of unsigned
Use size_t instead of unsigned to be consistent with the other methods.

This is an omission from Ifb24da16b3282b72a610dc8f6f6c98f37623edd5.

+++ src/lib/libk/bitmap.cpp
@@ -791,7 +791,7 @@ public:
    * \return 1 plus the index of the least significant set bit in the bitmap
    *         (starting at the given bit index).
    */
-  unsigned ffs(size_t bit) const
+  size_t ffs(size_t bit) const
   {
     unsigned long elem = _bits;
     elem >>= bit;
